utils added to Alert
